requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
pydantic==2.7.1
pydantic-settings==2.2.1
supabase==2.4.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.0
python-dotenv==1.0.1
fpdf2==2.7.9
python-multipart==0.0.9
psycopg2-binary==2.9.9
